Add s2n_cipher_preferences source files
Moves the definition of our cipher preferences from s2n_config.c to a
standalone module.
