+++ tls/s2n_cipher_suites.c
@@ -19,6 +19,7 @@
 
 #include "crypto/s2n_cipher.h"
 
+#include "tls/s2n_cipher_preferences.h"
 #include "tls/s2n_cipher_suites.h"
 #include "tls/s2n_connection.h"
 #include "tls/s2n_tls.h"
